Ra 01 first branch

+++ src/home/home.controller.ts
@@ -8,25 +8,32 @@ import {
   Post,
   Put,
   Query,
+  UnauthorizedException,
+  UseGuards,
 } from '@nestjs/common';
+import { UserType } from '@prisma/client';
 import { HomeService } from './home.service';
 import {
   CreateHomeDto,
   HomeResponseDto,
+  InquireDto,
   UpdateHomeDto,
 } from './dto/home.dto.ts/home.dto';
 import { PropertyType } from '@prisma/client';
+import { User, UserInfo } from 'src/user/decorators/user.decorator';
+import { AuthGuard } from '../guards/auth.guard';
+import { Roles } from 'src/decorators/roles.decorator';
 
 @Controller('home')
 export class HomeController {
   constructor(private readonly homeService: HomeService) {}
 
-  @Get()
+  @Get('/')
   getAllHomes() {
     return this.homeService.getAllHomes();
   }
 
-  @Get('')
+  @Get(':id')
   getHomeById(
     @Query('city') city?: string,
     @Query('minPrice') minPrice?: string,
@@ -50,22 +57,72 @@ export class HomeController {
     return this.homeService.getHomeById(filters);
   }
 
+  @Roles(UserType.REALTOR)
+  // @UseGuards(AuthGuard)
   @Post()
-  createHome(@Body() body: CreateHomeDto) {
-    return this.homeService.createHome(body);
+  createHome(@Body() body: CreateHomeDto, @User() user: UserInfo) {
+    // console.log(user);
+    return this.homeService.createHome(body, user.id);
+
+    // return user;
   }
+
+  @Roles(UserType.REALTOR)
+  // @UseGuards(AuthGuard)
   @Put(':id')
-  updateHome(
+  async updateHome(
     @Param('id', ParseIntPipe) id: number,
+    @User() user: UserInfo,
     @Body() body: UpdateHomeDto,
   ) {
+    const realtor = await this.homeService.getRealtorByHomeId(id);
+    if (realtor.id !== user.id) {
+      throw new UnauthorizedException();
+    }
     return this.homeService.updateHomeById(id, body);
   }
 
+  @Roles(UserType.REALTOR)
+  // @UseGuards(AuthGuard)
   @Delete(':id')
-  deleteHome(@Param('id', ParseIntPipe) id: number) {
+  async deleteHome(
+    @Param('id', ParseIntPipe) id: number,
+    @User() user: UserInfo,
+  ) {
+    const realtor = await this.homeService.getRealtorByHomeId(id);
+
+    if (realtor.id !== user.id) {
+      throw new UnauthorizedException('nao podes apagar!');
+    }
+    console.log('oh, não! apagámos a casa! foi-se :...X ');
     return this.homeService.deleteHouseById(id);
+  }
+
+  @Roles(UserType.BUYER)
+  @Post('/:id/inquire')
+  inquire(
+    @Param('id', ParseIntPipe) homeId: number,
+    @User() user: UserInfo,
+    @Body() { message }: InquireDto,
+  ) {
+    return this.homeService.inquire(user, homeId, message);
+  }
+
+  @Roles(UserType.REALTOR)
+  @Get('/:id/messages')
+  async getHomeMessages(
+    @Param('id', ParseIntPipe) id: number,
+    @User() user: UserInfo,
+  ) {
+    const realtor = await this.homeService.getRealtorByHomeId(id);
+
+    if (realtor.id !== user.id) {
+      throw new UnauthorizedException('nao podes apagar!');
+    }
+
+    return this.homeService.getMessagesByHome(id);
   }
 }
 
-// Parei aqui porque hoje (dia 26 de Maio) soube que vou para o Rotaro, outro projeto da Rad, e vou colocar isto em pausa!
+// 1- Buyer sends message to a realtor
+// 2- Realtor gets all messages
